+++ pkg/helm/factory.go
@@ -1,4 +1,4 @@
-package dependencies
+package helm
 
 import (
 	"context"
